changes to support ray tracing
