+++ src/shapes/obb.rs
@@ -1,102 +1,6 @@
-//! This module exports various shapes that can be used to intersect with
-use crate::aabb::AABB;
-use crate::bounding_hierarchy::IntersectionAABB;
-use crate::{Mat4, Point3, Quat, Real, Vector3};
+//! This module defines an Oriented Bounding Box and its intersection properties
+use crate::{Quat, Vector3, aabb::AABB, bounding_hierarchy::IntersectionAABB, Mat4};
 
-/// A representation of a Sphere
-pub struct Sphere {
-    /// Center of the sphere
-    pub center: Point3,
-    /// Radius of the sphere
-    pub radius: Real,
-}
-
-impl Sphere {
-    /// Creates a sphere centered on a given point with a radius
-    pub fn new(center: Point3, radius: Real) -> Sphere {
-        Sphere { center, radius }
-    }
-}
-
-impl IntersectionAABB for Sphere {
-    fn intersects_aabb(&self, aabb: &AABB) -> bool {
-        let vec = aabb.closest_point(self.center);
-        vec.distance_squared(self.center) < self.radius * self.radius
-    }
-}
-
-/// Representation of a capsule
-pub struct Capsule {
-    /// Start point of the line segment for the capsule
-    pub start: Point3,
-    /// Radius of the capsule
-    pub radius: Real,
-    /// Direction of the capsule's line segment
-    pub dir: Vector3,
-    /// Length of the capsules line segment
-    pub len: Real,
-}
-
-impl Capsule {
-    /// Creates a capsule given a start and end point for the center line and the radius around it
-    pub fn new(start: Point3, end: Point3, radius: Real) -> Capsule {
-        let line = end - start;
-        let dir = line.normalize();
-        let len = line.length();
-
-        Capsule {
-            start,
-            radius,
-            dir,
-            len,
-        }
-    }
-}
-
-impl IntersectionAABB for Capsule {
-    fn intersects_aabb(&self, aabb: &AABB) -> bool {
-        /*
-        // Use Distance from closest point
-        let mut point: Vector3 = self.start;
-        let mut curr_d = 0.0;
-        let max_sq = (self.len + self.radius) * (self.len + self.radius) ;
-        let r_sq = self.radius * self.radius;
-
-        loop {
-            let x = aabb.closest_point(point);
-            let d_sq = x.distance_squared(point);
-            println!("{:?} closest={:?} d={:?}", point, x, d_sq.sqrt());
-            if d_sq <= r_sq
-            {
-                return true;
-            }
-            if d_sq > max_sq || curr_d >= self.len
-            {
-                return false;
-            }
-            curr_d = (curr_d + d_sq.sqrt()).min(self.len);
-            point = self.start + (curr_d * self.dir);
-        }
-        */
-        let mut last = self.start;
-        loop {
-            let closest = &aabb.closest_point(last);
-            let center = nearest_point_on_line(&self.start, &self.dir, self.len, closest);
-            let sphere = Sphere {
-                center,
-                radius: self.radius,
-            };
-            if sphere.intersects_aabb(aabb) {
-                return true;
-            }
-            if last.distance_squared(center) < 0.0001 {
-                return false;
-            } else {
-                last = center;
-            }
-        }
-    }
-}
 
 /// Represents a box that can be rotated in any direction
 pub struct OBB {
@@ -264,70 +168,4 @@ fn back(matrix: Mat4) -> Vector3 {
 
 fn translation(matrix: Mat4) -> Vector3 {
     matrix.row(3).truncate().into()
-}
-
-/// Helper function that given a line segment and a target point, finds the closest point on the line segment to target
-pub fn nearest_point_on_line(start: &Point3, dir: &Vector3, len: Real, target: &Point3) -> Point3 {
-    let v = *target - *start;
-    let d = v.dot(*dir);
-    *start + (*dir * d.clamp(0.0, len))
-}
-
-#[cfg(test)]
-mod tests {
-    use crate::aabb::AABB;
-    use crate::bounding_hierarchy::IntersectionAABB;
-    use crate::shapes::{Capsule, OBB};
-    use crate::{Point3, Quat, Real, Vector3};
-
-    #[test]
-    fn basic_test_capsule() {
-        let min = Point3::new(0.0, 0.0, 0.0);
-        let max = Point3::new(1.0, 1.0, 1.0);
-        let aabb = AABB::empty().grow(&min).grow(&max);
-        let start = Point3::new(3.0, 0.0, 0.0);
-        let end = Point3::new(1.5, 0.0, 0.0);
-        let capsule = Capsule::new(start, end, 0.55);
-        assert!(capsule.intersects_aabb(&aabb));
-    }
-
-    #[test]
-    fn moving_test_capsule() {
-        let min = Point3::new(0.0, 0.0, 0.0);
-        let max = Point3::new(1.0, 1.0, 1.0);
-        let aabb = AABB::empty().grow(&min).grow(&max);
-        let start = Point3::new(0.5, 2.0, 2.0);
-        let end = Point3::new(0.5, 5.0, -1.0);
-        let capsule = Capsule::new(start, end, 1.45);
-        assert!(capsule.intersects_aabb(&aabb));
-
-        let dir = (start - end).normalize();
-        let offset: Real = 0.005;
-        println!("{}", dir);
-        for i in 0..800 {
-            println!("{}", i);
-            let pt = offset * dir * i as Real;
-            let cap = Capsule::new(start + pt, end + pt, 1.45);
-            assert!(cap.intersects_aabb(&aabb));
-        }
-    }
-
-    #[test]
-    fn basic_obb() {
-        let min = Point3::new(0.0, 0.0, 0.0);
-        let max = Point3::new(1.0, 1.0, 1.0);
-        let aabb = AABB::empty().grow(&min).grow(&max);
-
-        let ori = Quat::from_axis_angle(Vector3::new(1.0, 0.0, 0.0).into(), 0.785398);
-        let extents = Vector3::new(0.5, 0.5, 0.5);
-        let pos = Vector3::new(0.5, 2.2, 0.5);
-
-        let obb = OBB {
-            orientation: ori,
-            extents,
-            center: pos,
-        };
-
-        assert!(obb.intersects_aabb(&aabb));
-    }
 }
