+++ src/shapes/sphere.rs
@@ -1,0 +1,53 @@
+//! This module defines a Sphere and it's intersection algorithms
+use crate::{Point3, Real, bounding_hierarchy::IntersectionAABB, aabb::AABB, ray::{IntersectionRay, Ray, Intersection}, Vector3};
+
+
+/// A representation of a Sphere
+pub struct Sphere {
+    /// Center of the sphere
+    pub center: Point3,
+    /// Radius of the sphere
+    pub radius: Real,
+}
+
+impl Sphere {
+    /// Creates a sphere centered on a given point with a radius
+    pub fn new(center: Point3, radius: Real) -> Sphere {
+        Sphere { center, radius }
+    }
+}
+
+impl IntersectionAABB for Sphere {
+    fn intersects_aabb(&self, aabb: &AABB) -> bool {
+        let vec = aabb.closest_point(self.center);
+        vec.distance_squared(self.center) < self.radius * self.radius
+    }
+}
+
+impl IntersectionRay for Sphere {
+    fn intersects_ray(&self, ray: &Ray, t_min: Real, t_max: Real) -> Option<Intersection> {
+        let oc = ray.origin - self.center;
+        let a = ray.direction.length_squared();
+        let half_b = oc.dot(ray.direction);
+        let c = oc.length_squared() - self.radius * self.radius;
+        let discriminant = half_b*half_b - a*c;
+
+        if discriminant < 0. {
+            return None;
+        }
+
+        let sqrtd = discriminant.sqrt();
+
+        let mut toi = (-half_b - sqrtd) / a;
+        if toi < t_min || t_max < toi {
+            toi = (-half_b + sqrtd) / a;
+            if toi < t_min || t_max < toi {
+                return None
+            }
+        }
+
+        let out_norm = (ray.at(toi) - self.center).normalize();
+        let (norm, back_face) = ray.face_normal(out_norm);
+        Some(Intersection::new(toi, 0., 0., norm, back_face))
+    }
+}
